last_version

+++ index.js
@@ -1,9 +1,12 @@
 const express = require("express");
 const { initDB } = require("./db");
-const ToDo = require("./db/models/ToDo model");
+const ToDo = require("./db/models/ToDo.model");
+const User = require("./db/models/User.model");
+const Token = require("./db/models/Token.model");
 const app = express();
 const port = 3100;
 const cors = require("cors");
+const { nanoid } = require("nanoid");
 
 app.use(express.json());
 app.use(cors());
@@ -12,99 +15,97 @@ app.listen(port, () => {
   console.log("Application listening on port " + port);
 });
 
-app.get("/sum", (req, res) => {
-  const a = req.body.a;
-  const b = req.body.b;
-
-  if (a === String(a) || b === String(b)) {
-    res.json({ message: "ERROR" });
-  } else {
-    const sum = a + b;
-    res.json({ sum: sum });
-  }
-});
-
-app.post("/reverse-case", (req, res) => {
-  const text_str = req.body.text;
-  let new_str = "";
-  for (let i = 0; i < text_str.length - 1; i++) {
-    let ch = text_str.charAt(i);
-    if (ch === ch.toUpperCase()) {
-      ch = ch.toLowerCase();
-    } else {
-      ch = ch.toUpperCase();
-    }
-    new_str += ch;
-  }
-  res.json({ new_str: new_str });
-});
-
-app.put("/obj-to-array", (req, res) => {
-  const x = req.body.x;
-  const massKeys = Object.keys(x);
-  const massValues = Object.values(x);
-  console.log(massKeys, massValues);
-  let mass = [];
-  for (let i = 0; i < massKeys.length; i++) {
-    mass.push({
-      key: massKeys[i],
-      Value: massValues[i],
-    });
-  }
-  res.json({ mass: mass });
-});
-
-app.patch("/reverse-array", (req, res) => {
-  const arr = req.body.arr;
-  let reverse_arr = [];
-  let n = 0;
-  for (let i = arr.length - 1; i >= 0; i--) {
-    reverse_arr[n] = arr[i];
-    n += 1;
-  }
-  res.json({ reverse_arr: reverse_arr });
-});
-
-app.delete("/delete", (req, res) => {
-  mass = req.body.mass;
-  const result = Array.from(new Set(mass));
-  res.json({ newmass: result });
-});
-
 initDB();
 
-app.get("/api/todo", async (req, res) => {
+app.post("/registration", async (req, res) => {
   try {
-    const list = await ToDo.findAll();
+    if ((await User.findOne({ where: { mail: req.body.mail } })) == null) {
+      const single_user = await User.create({
+        password: req.body.password,
+        mail: req.body.mail,
+        firstName: req.body.firstName,
+        lastName: req.body.lastName,
+      });
+      res.json({ user: single_user });
+    } else {
+      res.status(400).json({ message: "EMail уже существует" });
+    }
+  } catch (error) {
+    res.status(500).json({ message: "ERROR", error });
+  }
+});
+
+app.post("/auth", async (req, res) => {
+  try {
+    const tokenValue = nanoid(64);
+    const user = await User.findOne({
+      where: { mail: req.body.mail, password: req.body.password },
+    });
+    if (user != null) {
+      const token = await Token.create({
+        value: tokenValue,
+        userId: user.id,
+      });
+      res.json({ token: token });
+    } else {
+      res.status(400).json({ message: "НЕПРАВИЛЬНЫЕ ДАННЫЕ" });
+    }
+  } catch (error) {
+    res.status(500).json({ message: "ERROR", error });
+  }
+});
+
+app.use(async (req, res, next) => {
+  if (req.headers.value === null) {
+    return res.status(400).json({ message: "troken not provided" });
+  }
+  const token = await Token.findOne({
+    where: { value: req.headers.value },
+  }); 
+  if (token === null) {
+    return res.status(400).json({ message: "troken not found" });
+  }
+  req.userId = token.userId;
+  next();
+});
+
+app.get("/api/todos", async (req, res) => {
+  try {
+    console.log("get", req.headers);
+    const list = await ToDo.findAll({
+      where: {
+        userId: req.userId,
+      },
+    });
     res.json({ todoList: list });
   } catch (error) {
     res.status(500).json({ message: "ERROR", error });
   }
 });
 
-app.get("/api/todo/:id", async (req, res) => {
-  try {
-    //как сократить запись
-    const single_todo = await ToDo.findOne({
-      where: {
-        id: req.params.id,
-      },
-    });
-    if (single_todo === null) {
-      res.status(404).json({ message: "ToDo not found" });
-    } else {
-      res.json({ todo: single_todo });
-    }
-  } catch (error) {
-    res.status(500).json({ message: "ERROR" });
-  }
-});
+// app.get("/api/todo/:id", async (req, res) => {
+//   try {
+//     const single_todo = await ToDo.findOne({
+//       where: {
+//         id: req.params.id,
+//       },
+//     });
+//     if (single_todo === null) {
+//       res.status(404).json({ message: "ToDo not found" });
+//     } else {
+//       res.json({ todo: single_todo });
+//     }
+//   } catch (error) {
+//     res.status(500).json({ message: "ERROR" });
+//   }
+// });
 
-app.post("/api/todo", async (req, res) => {
+app.post("/api/todos", async (req, res) => {
   try {
     const todo = await ToDo.create({
       title: req.body.title,
       description: req.body.description,
+      userId: req.userId,
     });
     res.json({ todo: todo });
   } catch (error) {
@@ -112,12 +113,13 @@ app.post("/api/todo", async (req, res) => {
   }
 });
 
-app.patch("/api/todo/:id", async (req, res) => {
+app.patch("/api/todos/:id", async (req, res) => {
   try {
     if (
       (await ToDo.findOne({
         where: {
           id: req.params.id,
+          userId: req.userId,
         },
       })) == null
     ) {
@@ -131,6 +133,7 @@ app.patch("/api/todo/:id", async (req, res) => {
         {
           where: {
             id: req.params.id,
+            userId: req.userId,
           },
         }
       );
@@ -141,12 +144,13 @@ app.patch("/api/todo/:id", async (req, res) => {
   }
 });
 
-app.delete("/api/todo/:id", async (req, res) => {
+app.delete("/api/todos/:id", async (req, res) => {
   try {
     if (
       (await ToDo.findOne({
         where: {
           id: req.params.id,
+          userId: req.userId,
         },
       })) == null
     ) {
@@ -155,6 +159,7 @@ app.delete("/api/todo/:id", async (req, res) => {
       await ToDo.destroy({
         where: {
           id: req.params.id,
+          userId: req.userId
         },
       });
       res.json({ message: "Complete" });
@@ -164,11 +169,38 @@ app.delete("/api/todo/:id", async (req, res) => {
   }
 });
 
-app.delete("/api/todo", async (req, res) => {
+app.delete("/api/todos", async (req, res) => {
   try {
-    await ToDo.destroy({ where: {} });
+    await ToDo.destroy({ where: { userId: req.userId } });
     res.json({ message: "Complete" });
   } catch (error) {
     res.status(500).json({ message: "ERROR" });
   }
 });
+
+app.delete("/logout/", async (req, res) => {
+  try {
+    await Token.destroy({ where: { userId: req.userId } });
+    res.json({ message: "Complete" });
+  } catch (error) {
+    res.status(500).json({ message: "ERROR" });
+  }
+});
+
+// app.get("/getTokens", async(req, res) => {
+//   try {
+//     const list = await Token.findAll();
+//     res.json({ tokenList: list });
+//   } catch (error) {
+//     res.status(500).json({ message: "ERROR", error });
+//   }
+// })
+
+// app.get("/getUsers", async(req, res) => {
+//   try {
+//     const list = await User.findAll();
+//     res.json({ UserList: list });
+//   } catch (error) {
+//     res.status(500).json({ message: "ERROR", error });
+//   }
+// })
